// src/data/enums/portrait/enumQuestionsPortrait.ts
export enum UserPortraitQuestions {
  GENDER_IDENTITY = 'Как вы себя определяете (ваша гендерная идентичность)?',
  TRANSGENDER_IDENTITY = 'Идентифицируете ли вы себя как трансгендерного человека? ',
  SEXUAL_ORIENTATION = 'Ваша сексуальная ориентация',
  AGE = 'Ваш возраст',
  RESIDENCE = 'Где вы проживаете? ',
  REGION = 'Выберите, пожалуйста, регион, в котором вы проживаете',
  FEDERAL_DISTRICT = 'Федеральный округ',
  FRIENDS_AWARE_OF_LGBT = 'Кто из ваших друзей знает о том, что вы ЛГБТК+ персона?',
  FAMILY_AWARE_OF_LGBT = 'Кто из ваших родных знает о том, что вы ЛГБТК+ персона?',
  CLASSMATES_COLLEAGUES_AWARE_OF_LGBT = 'Кто из ваших одноклассников/однокурсников/коллег (в нынешнем учебном/рабочем коллективе)'
      + ' знает о том, что вы ЛГБТК+ персона?',
  SOCIAL_CIRCLE_LGBT = 'Ваш круг общения, в основном, состоит из...',
  SOCIAL_CIRCLE_ATTITUDE = 'Ваш круг общения – это, в основном, (по вашим ощущениям)…',
  EDUCATION = 'Ваше образование',
}