// src/data/enums/discrimination/enumQuestionsDiscrimination.ts
// Учеба/работа
export enum EmploymentQuestions {
  PRESSURE_DUE_TO_IDENTITY = 'Сталкивались ли вы с каким-либо давлением на работе/учёбе из-за сексуальной ориентации и/или трансгендерности в 2022 году?',
  JOB_OR_EDUCATION_LOSS = 'Приходилось ли вам в 2022 году терять работу/место учёбы из-за сексуальной ориентации и/или трансгендерности?',
}

// Насилие
export enum ViolenceQuestions {
  PHYSICAL_VIOLENCE = 'Подвергались ли вы в 2022 году физическому насилию из-за сексуальной ориентации и/или трансгендерности?',
  SEXUAL_VIOLENCE = 'Подвергались ли вы в 2022 году сексуализированному насилию (в том числе домогательствам и попыткам изнасилования) из-за вашей сексуальной ориентации и/или трансгендерности?',
  THREATS_OF_PHYSICAL_VIOLENCE = 'Сталкивались ли вы в 2022 году с угрозами физическим насилием из-за сексуальной ориентации и/или трансгендерности?',
  DOMESTIC_VIOLENCE = 'Сталкивались ли вы в 2022 году с домашним насилием из-за сексуальной ориентации и/или трансгендерности (со стороны родителей, родственников, супругов, партнёров)?',
  PROPERTY_DAMAGE = 'Сталкивались ли вы в 2022 году со случаями причинения вреда вашему имуществу, уничтожения, кражи/хищения имущества из-за вашей сексуальной ориентации и/или трансгендерности?',
  ROBBERY = 'Сталкивались ли вы в 2022 году с грабежом по типу “подставных свиданий”?',
  BLACKMAIL = 'Сталкивались ли вы в 2022 году с шантажом/вымогательством с угрозой раскрытия информации о вашей сексуальной ориентации или трансгендерности?',
  ILLEGAL_USE_OF_PERSONAL_DATA = 'Сталкивались ли вы в 2022 году со случаями незаконного использования ваших персональных данных из-за вашей сексуальной ориентации и/или трансгендерности?',
  ONLINE_HARASSMENT = 'Сталкивались ли именно вы в 2022 году с преследованиями в интернете из-за вашей сексуальной ориентации и/или трансгендерности?',
  UNLAWFUL_DETENTION_OR_ARREST = 'Сталкивались ли вы в 2022 году с незаконным задержанием или арестом из-за вашей сексуальной ориентации и/или трансгендерности?',
  CRIMES_DUE_TO_HOMOPHOBIA = 'Сталкивались ли вы в 2022 году с преступлениями на почве гомофобии/трансфобии (нападения, насилия, угроз и т.п.)? И обращались ли вы при этом в полицию?',
}