// src/data/enums/discrimination/enumAnswersDiscrimination.ts
export enum ViolenceAnswers {
  NO_ENCOUNTER_WOULD_REPORT_TO_POLICE = 'Не сталкивались, но если бы столкнулись, то обратились бы в полицию',
  ENCOUNTERED_AND_REPORTED_TO_POLICE = 'Сталкивались и обращались в полицию',
  ENCOUNTERED_BUT_DID_NOT_REPORT_TO_POLICE = 'Сталкивались, но не обращались в полицию',
  NO_ENCOUNTER_WOULD_NOT_REPORT_TO_POLICE = 'Не сталкивались, но если бы столкнулись, то не стали бы обращаться в полицию',
}